email/add_ecr/requirements.txt
psycopg2-binary
python-dotenv
pandas
boto3